requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.13.0
    # via tensorboard
argon2-cffi==20.1.0
    # via notebook
attrs==21.2.0
    # via jsonschema
backcall==0.2.0
    # via ipython
bleach==4.0.0
    # via nbconvert
bravado==11.0.3
    # via neptune-client
bravado-core==5.17.0
    # via bravado
cachetools==4.2.2
    # via google-auth
git+https://gitlab.cern.ch/mmajewsk/calina
    # via -r requirements.in
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via argon2-cffi
charset-normalizer==2.0.4
    # via requests
click==8.0.1
    # via neptune-client
cycler==0.10.0
    # via matplotlib
debugpy==1.4.1
    # via ipykernel
decorator==5.0.9
    # via ipython
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.3
    # via
    #   jupyter-client
    #   nbconvert
fsspec==2021.7.0
    # via pytorch-lightning
future==0.18.2
    # via
    #   neptune-client
    #   pytorch-lightning
    #   torch
gitdb==4.0.7
    # via gitpython
gitpython==3.1.18
    # via neptune-client
google-auth==1.35.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.5
    # via tensorboard
grpcio==1.39.0
    # via tensorboard
idna==3.2
    # via
    #   jsonschema
    #   requests
ipykernel==6.2.0
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==7.26.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   nbformat
    #   notebook
    #   qtconsole
    #   traitlets
ipywidgets==7.6.3
    # via jupyter
jedi==0.18.0
    # via ipython
jinja2==3.0.1
    # via
    #   nbconvert
    #   notebook
joblib==1.0.1
    # via scikit-learn
jsonpointer==2.1
    # via jsonschema
jsonref==0.2
    # via bravado-core
jsonschema[format]==3.2.0
    # via
    #   bravado-core
    #   nbformat
    #   swagger-spec-validator
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==7.0.1
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.0
    # via jupyter
jupyter-core==4.7.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.0
    # via ipywidgets
kiwisolver==1.3.1
    # via matplotlib
markdown==3.3.4
    # via tensorboard
markupsafe==2.0.1
    # via jinja2
matplotlib==3.4.3
    # via -r requirements.in
matplotlib-inline==0.1.2
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
monotonic==1.6
    # via bravado
msgpack==1.0.2
    # via
    #   bravado
    #   bravado-core
nbclient==0.5.4
    # via nbconvert
nbconvert==6.1.0
    # via
    #   jupyter
    #   notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
neptune-client==0.10.7
    # via -r requirements.in
nest-asyncio==1.5.1
    # via
    #   jupyter-client
    #   nbclient
notebook==6.4.3
    # via
    #   jupyter
    #   widgetsnbextension
numpy==1.21.2
    # via
    #   calina-dataset
    #   matplotlib
    #   pandas
    #   pytorch-lightning
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   torch
oauthlib==3.1.1
    # via
    #   neptune-client
    #   requests-oauthlib
packaging==21.0
    # via
    #   bleach
    #   neptune-client
pandas==1.3.2
    # via
    #   -r requirements.in
    #   calina-dataset
    #   neptune-client
pandocfilters==1.4.3
    # via nbconvert
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.1
    # via
    #   matplotlib
    #   neptune-client
plotly==5.2.2
    # via -r requirements.in
prometheus-client==0.11.0
    # via notebook
prompt-toolkit==3.0.20
    # via
    #   ipython
    #   jupyter-console
protobuf==3.17.3
    # via tensorboard
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.20
    # via cffi
pygments==2.10.0
    # via
    #   ipython
    #   jupyter-console
    #   jupyterlab-pygments
    #   nbconvert
    #   qtconsole
pyjwt==2.1.0
    # via neptune-client
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   bravado
    #   bravado-core
    #   jupyter-client
    #   matplotlib
    #   pandas
pytorch-lightning==1.1.0
    # via -r requirements.in
pytz==2021.1
    # via
    #   bravado-core
    #   pandas
pyyaml==5.4.1
    # via
    #   bravado
    #   bravado-core
    #   pytorch-lightning
    #   swagger-spec-validator
pyzmq==22.2.1
    # via
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.1.1
    # via jupyter
qtpy==1.10.0
    # via qtconsole
requests==2.26.0
    # via
    #   bravado
    #   neptune-client
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.0
    # via
    #   google-auth-oauthlib
    #   neptune-client
rfc3987==1.3.8
    # via jsonschema
rsa==4.7.2
    # via google-auth
scikit-learn==0.24.2
    # via sklearn
scipy==1.7.1
    # via scikit-learn
send2trash==1.8.0
    # via notebook
simplejson==3.17.5
    # via
    #   bravado
    #   bravado-core
six==1.16.0
    # via
    #   absl-py
    #   argon2-cffi
    #   bleach
    #   bravado
    #   bravado-core
    #   cycler
    #   google-auth
    #   grpcio
    #   jsonschema
    #   neptune-client
    #   plotly
    #   protobuf
    #   python-dateutil
    #   swagger-spec-validator
sklearn==0.0
    # via -r requirements.in
smmap==4.0.0
    # via gitdb
strict-rfc3339==0.7
    # via jsonschema
swagger-spec-validator==2.7.3
    # via bravado-core
tenacity==8.0.1
    # via plotly
tensorboard==2.6.0
    # via pytorch-lightning
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.0
    # via tensorboard
terminado==0.11.1
    # via notebook
testpath==0.5.0
    # via nbconvert
threadpoolctl==2.2.0
    # via scikit-learn
torch==1.6.0
    # via
    #   -r requirements.in
    #   pytorch-lightning
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.62.2
    # via
    #   calina-dataset
    #   pytorch-lightning
traitlets==5.0.5
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typing-extensions==3.10.0.0
    # via bravado
urllib3==1.26.6
    # via
    #   neptune-client
    #   requests
wcwidth==0.2.5
    # via prompt-toolkit
webcolors==1.11.1
    # via jsonschema
webencodings==0.5.1
    # via bleach
websocket-client==1.2.1
    # via neptune-client
werkzeug==2.0.1
    # via tensorboard
wheel==0.37.0
    # via tensorboard
widgetsnbextension==3.5.1
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
